move convertJsontoarray from exportView.ts to store.ts
