+++ src/store/store.ts
@@ -11,9 +11,9 @@ import display, { DisplayState } from './display';
 import filter, { FilterStateJSON, TimeSlot } from './filter';
 import modal from './modal';
 import noti from './notification';
-import palette, { PaletteState } from './palette';
+import palette, { PaletteState, Palette } from './palette';
 import profile from './profile';
-import schedule, { ScheduleStateJSON } from './schedule';
+import schedule, { ScheduleStateJSON, ScheduleStore } from './schedule';
 import semester, { SemesterState } from './semester';
 import status from './status';
 import Expirable from '@/data/Expirable';
@@ -368,5 +368,96 @@ export default class Store extends Vue {
         }
         await this.loadProfile();
         this.status.loading = false;
+    }
+
+    /**
+     * See [[App.parseFromURL]]
+     * convert JSON string to tuple of tuples to reduce the num of chars
+     * @author Zichao Hu
+     * @param jsonString
+     */
+    convertJsonToArray(jsonString: string) {
+        /*
+        result =>
+        name
+        modified
+        currentSemester.id
+        currentSemester.name,
+        _earliest
+        _fullHeight
+        _lastest
+        _maxNumSchedules
+        _numsearchResults
+        _partialHeight
+        binary: combineSections, enableFuzzy, enableLog, expandOnEntering, multiSelect,
+        showClasslistTitle, showInstructor, showRoom, showTime, standard
+        timeSlots
+        binary: allowClosed, Waitlist, mode
+        sortOptions: name_initial ascii code :[c,d,l,n,v,I] --> could change in order
+        binary: enabled/reverse
+        schedule
+        palette
+        */
+
+        // get values from the json object
+        const json: SemesterStorage = JSON.parse(jsonString);
+        const { name, modified, currentSemester, display, filter, schedule, palette } = json;
+
+        // add first four value the the array
+        const result = [];
+        result.push(name, modified, currentSemester.id, currentSemester.name);
+
+        // compressing display
+        // get all keys in the display object and sort them
+        const display_keys = Object.keys(display).sort();
+
+        // convert to binary, the first key => the first/rightmost bit
+        // there are 10 keys to consider
+        let display_bit = 0;
+        let counter = 1;
+        for (const key of display_keys) {
+            if (display[key] === true) {
+                display_bit |= counter;
+                counter <<= 1;
+            } else if (display[key] === false) {
+                counter <<= 1;
+            } else {
+                result.push(display[key]);
+            }
+        }
+        result.push(display_bit);
+
+        // compressing filter
+        // add timeSlots to array
+        result.push(filter.timeSlots);
+
+        // convert allowClosed, allowWaitlist, mode to binary
+        let filter_bit = 0;
+        if (filter.allowClosed) filter_bit += 2 ** 0;
+        if (filter.allowWaitlist) filter_bit += 2 ** 1;
+        if (filter.sortOptions.mode === 1) filter_bit += 2 ** 2;
+
+        result.push(filter_bit);
+
+        // sorting
+        // add all initial ascii to the array in order
+        // add the binary of their respective state: enabled or reverse
+        counter = 1;
+        filter_bit = 0;
+        for (const sortBy of filter.sortOptions.sortBy) {
+            const ascii = sortBy.name.charCodeAt(0);
+            result.push(ascii);
+            if (sortBy.enabled) filter_bit |= counter;
+            counter <<= 1;
+            if (sortBy.reverse) filter_bit |= counter;
+            counter <<= 1;
+        }
+        result.push(filter_bit);
+
+        // add schedule and palette objects to the array
+        result.push(ScheduleStore.compressJSON(schedule));
+        result.push(Palette.compressJSON(palette));
+        console.log(result);
+        return JSON.stringify(result);
     }
 }
